Tuple with generics done
