+++ RayCasterEngine/Program.cs
@@ -7,14 +7,14 @@ namespace RayCasterEngine
         static void Main(string[] args)
         {   
             var projectilePosition = Tuple3d.Point(0, 1, 0);
-            var projectileVelocity = Tuple3d.Normalize(Tuple3d.Vector(1, 1, 0));
+            var projectileVelocity = RayCasterEngine.Tuple<Tuple3d>.Normalize(Tuple3d.Vector(1, 1, 0));
             var projectile = new Projectile(projectilePosition, projectileVelocity);
 
             var envGravity = Tuple3d.Vector(0, -0.1, 0);
             var envWind = Tuple3d.Vector(-0.01, 0, 0);
             var env = new RayCasterEngine.Environment(envGravity, envWind);
 
-            while (projectile.Position.Y >= 0)
+            while (projectile.Position.Data.Y >= 0)
             {
                 projectile = Projectile.Tick(env, projectile);
                 System.Console.WriteLine(projectile.GetPosition());
