+++ RayCasterTest/TupleTests.fs
@@ -24,7 +24,7 @@ let ``A tuple with w=0 is a vector`` () =
 let ``Point() creates tuples with w=1`` () =
     let x, y, z = (4.0, -4.0, 3.0)
     let point = Tuple3d.Point (x, y, z)
-    let expected = Tuple3d(x, y, z, 1.0)
+    let expected = Tuple (Tuple3d(x, y, z, 1.0))
 
     point = expected
     |> Assert.True
@@ -33,16 +33,16 @@ let ``Point() creates tuples with w=1`` () =
 let ``Vector() creates tuples with w=0`` () =
     let x, y, z = (4.0, -4.0, 3.0)
     let vector = Tuple3d.Vector (x, y, z)
-    let expected = Tuple3d(x, y, z, 0.0)
+    let expected = Tuple (Tuple3d(x, y, z, 0.0))
 
     vector = expected
     |> Assert.True
 
 [<Fact>]
 let ``Adding two tuples`` () =
-    let tuple1 = Tuple3d (3.0, -2.0, 5.0, 1.0)
-    let tuple2 = Tuple3d (-2.0, 3.0, 1.0, 0.0)
-    let expected = Tuple3d (1.0, 1.0, 6.0, 1.0)
+    let tuple1 = Tuple (Tuple3d (3.0, -2.0, 5.0, 1.0))
+    let tuple2 = Tuple (Tuple3d (-2.0, 3.0, 1.0, 0.0))
+    let expected = Tuple (Tuple3d (1.0, 1.0, 6.0, 1.0))
 
     tuple1 + tuple2 = expected
     |> Assert.True
@@ -73,37 +73,37 @@ let ``Subtracting two vectors`` () =
 
 [<Fact>]
 let ``Negating a tuple`` () =
-    let t = Tuple3d (1.0, -2.0, 3.0, -4.0)
-    let expected = Tuple3d (-1.0, 2.0, -3.0, 4.0)
+    let t = Tuple (Tuple3d (1.0, -2.0, 3.0, -4.0))
+    let expected = Tuple (Tuple3d (-1.0, 2.0, -3.0, 4.0))
 
     -t = expected
     |> Assert.True
 
 [<Fact>]
 let ``Multiplying a tuple by a scalar`` () =
-    let a = Tuple3d (1.0, -2.0, 3.0, -4.0)
+    let a = Tuple (Tuple3d (1.0, -2.0, 3.0, -4.0))
 
-    a * 3.5 = Tuple3d (3.5, -7.0, 10.5, -14.0)
+    a * 3.5 = Tuple (Tuple3d (3.5, -7.0, 10.5, -14.0))
     |> Assert.True
 
 [<Fact>]
 let ``Multiplying a tuple by a fraction`` () =
-    let a = Tuple3d (1.0, -2.0, 3.0, -4.0)
+    let a = Tuple (Tuple3d (1.0, -2.0, 3.0, -4.0))
     
-    a * 0.5 = Tuple3d (0.5, -1.0, 1.5, -2.0)
+    a * 0.5 = Tuple (Tuple3d (0.5, -1.0, 1.5, -2.0))
     |> Assert.True
 
 [<Fact>]
 let ``Dividing a tuple by a scalar`` () =
-    let a = Tuple3d (1.0, -2.0, 3.0, -4.0)
+    let a = Tuple (Tuple3d (1.0, -2.0, 3.0, -4.0))
 
-    a / 2.0 = Tuple3d (0.5, -1.0, 1.5, -2.0)
+    a / 2.0 = Tuple (Tuple3d (0.5, -1.0, 1.5, -2.0))
     |> Assert.True
 
 [<Fact>]
 let ``Computing the magnutide of vector(1, 0, 0)`` () =
     let v = Tuple3d.Vector (1.0, 0.0, 0.0)
-    let magnitudeV = Tuple3d.Magnitude (v)
+    let magnitudeV = Tuple.Magnitude(v);
 
     magnitudeV = 1.0
     |> Assert.True
@@ -111,7 +111,7 @@ let ``Computing the magnutide of vector(1, 0, 0)`` () =
 [<Fact>]
 let ``Computing the magnitude of vector(0, 1, 0)`` () =
     let v = Tuple3d.Vector (0.0, 1.0, 0.0)
-    let magnitudeV = Tuple3d.Magnitude (v)
+    let magnitudeV = Tuple.Magnitude (v)
 
     magnitudeV = 1.0
     |> Assert.True
@@ -119,7 +119,7 @@ let ``Computing the magnitude of vector(0, 1, 0)`` () =
 [<Fact>]
 let ``Computing the magnitude of vector(0, 0, 1)`` () =
     let v = Tuple3d.Vector (0.0, 0.0, 1.0)
-    let magnitudeV = Tuple3d.Magnitude (v)
+    let magnitudeV = Tuple.Magnitude (v)
 
     magnitudeV = 1.0
     |> Assert.True
@@ -127,7 +127,7 @@ let ``Computing the magnitude of vector(0, 0, 1)`` () =
 [<Fact>]
 let ``Computing the magnitude of vector(1, 2, 3)`` () =
     let v = Tuple3d.Vector (1.0, 2.0, 3.0)
-    let magnitudeV = Tuple3d.Magnitude (v)
+    let magnitudeV = Tuple.Magnitude (v)
 
     magnitudeV = sqrt 14.0
     |> Assert.True
@@ -135,7 +135,7 @@ let ``Computing the magnitude of vector(1, 2, 3)`` () =
 [<Fact>]
 let ``Computing the magnitude of vector(-1, -2, -3)`` () =
     let v = Tuple3d.Vector (-1.0, -2.0, -3.0)
-    let magnitudeV = Tuple3d.Magnitude (v)
+    let magnitudeV = Tuple.Magnitude (v)
 
     magnitudeV = sqrt 14.0
     |> Assert.True
@@ -143,7 +143,7 @@ let ``Computing the magnitude of vector(-1, -2, -3)`` () =
 [<Fact>]
 let ``Normalizing vector(4, 0, 0) gives (1, 0, 0)`` () =
     let v = Tuple3d.Vector (4.0, 0.0, 0.0)
-    let normalV = Tuple3d.Normalize(v)
+    let normalV = Tuple.Normalize(v)
 
     normalV = Tuple3d.Vector (1.0, 0.0, 0.0)
     |> Assert.True
@@ -151,7 +151,7 @@ let ``Normalizing vector(4, 0, 0) gives (1, 0, 0)`` () =
 [<Fact>]
 let ``Normalizing vector(1, 2, 3)`` () =
     let v = Tuple3d.Vector (1.0, 2.0, 3.0)
-    let normalV = Tuple3d.Normalize(v)
+    let normalV = Tuple.Normalize(v)
 
     normalV = Tuple3d.Vector (0.26726, 0.53452, 0.80178)
     |> Assert.True
@@ -159,16 +159,16 @@ let ``Normalizing vector(1, 2, 3)`` () =
 [<Fact>]
 let ``The magnitude of a normalized vector`` () =
     let v = Tuple3d.Vector (1.0, 2.0, 3.0)
-    let normalV = Tuple3d.Normalize(v)
+    let normalV = Tuple.Normalize(v)
 
-    Tuple3d.Magnitude (normalV) = 1.0
+    Tuple.Magnitude (normalV) = 1.0
     |> Assert.True
 
 [<Fact>]
 let ``The dot product of two tuples`` () =
     let a = Tuple3d.Vector (1.0, 2.0, 3.0)
     let b = Tuple3d.Vector (2.0, 3.0, 4.0)
-    let dotProduct = Tuple3d.DotProduct(a, b)
+    let dotProduct = Tuple.DotProduct(a, b)
 
     dotProduct = 20.0
     |> Assert.True
@@ -177,8 +177,8 @@ let ``The dot product of two tuples`` () =
 let ``The cross product of two vectors`` () =
     let a = Tuple3d.Vector (1.0, 2.0, 3.0)
     let b = Tuple3d.Vector (2.0, 3.0, 4.0)
-    let crossProductAB = Tuple3d.CrossProduct(a, b)
-    let crossProductBA = Tuple3d.CrossProduct(b, a)
+    let crossProductAB = Tuple3d.CrossProduct(a.Data, b.Data)
+    let crossProductBA = Tuple3d.CrossProduct(b.Data, a.Data)
 
     crossProductAB = Tuple3d.Vector (-1.0, 2.0, -1.0)
     |> (&&) (crossProductBA = Tuple3d.Vector (1.0, -2.0, 1.0))
